fbdev: msm: somc_panel: Fix onoff state and VPS cmd at poweroff
The onoff state has to be updated even if we don't actually
send any off commands to the display because, to power if off,
we may also just cut power completely, which will also reset
the DriverIC completely.

Also, we were already checking for DSI master panel, but
obviously the FPS (and vsync) handler should not even be
touched for a slave panel, since it automatically uses the
master panel configuration.
This is obviously for dual-dsi panels, not for dual panels.