arm: DT: dsi-panel-somc: Update all SoMC panels with TE height
This is more a hack than something completely proper, but at the
same time it's .. proper.
Explaining, we had a hack in the mdss_mdp_intf_cmd driver that
was actually doing exactly what I just configured in the DTs
which enabled us to use the FPS Manager in Dynamic Mode, instead
of the onoff mode, allowing us to get better scaling of the
configured FPS, hence slightly more battery life, with the
scaling itself happening seamlessly, exactly when needed, and
without any slowdown whatsoever.
Now, code hacks are bad.... and since we can do the same in DT
it's just better to.

Signed-off-by: Erik Castricum <git@cas-online.nl>
Change-Id: If2c4a6ce39cb72769a9b11d0b97ab7f7f3d38548