video: fbdev: msm: pp: Fix for kernel-driven PCC/PA corrections
These features are originally designed to get the adjustments
copied from userspace but, with somc_panel, we are doing it
entirely in kernel space.
If the copy_from_user call fails we can check kernel config
payload and, if it's set, we can continue setting up the color
adjustments just fine.

This will not break the userspace correction logic.