media: msm: camera_v2: cpp: Add compatibility with MSM8916-class GDSCR
The MSM8916-class GDSCR needs to be controlled directly by
IOW from the cpp driver.
Separate the GDSCR status update functions and refer to the
new one as "GDSCRv2" and the old one as "GDSCRv1".

Please note that the v1/v2 naming convention is done for
a pure cosmetic and practical reason: in reality, there is
no indication anywhere of the existance of those precise
versions.