mmc: sdio: Add new capability for nonstandard nonremovable cards
There are some SDIO cards that don't support plug in/out in
a standard way: once they are plugged in, it will not be
possible to plug them out and retain functionality through
suspend and resume.
This patch flags the card as NONREMOVABLE only after it has
been detected (otherwise, the detection handler will not
work).

In the example case of a BCM4356 SDIO card, the system may
start with no/unpowered card. We powerup the card and the
kernel recognizes it, then the specific driver for the chip
loads a firmware into its RAM: starting from this moment,
if we power down the card, we will not be able to power it
up again, since it will lose the firmware stored in RAM
and will never boot up anymore.

To make things worst, the whole BCM chip may be powered
down when the SDIO card gets a power cut, producing a
functionality breakage: the chip will be disconnected from
the WiFi network and the connection loss will be permanent,
as no sdiowakeup interrupt will fire: the chip is down.

Such cards do _not_ support any change in voltage once the
card setup is done, otherwise Bad Things Happen (TM), hence
the only way to get them working properly is to suspend the
host controller while keeping the card completely alive and
relying on its wireless driver to set it to low power mode.

Please note that, about power consumption, mileage may vary:
some wireless chips support a very very low power mode with
connection keepalive, some others do not, but this case is
not covered in this patch and will never be in this place,
as this is intended to only address problems coming from
the Linux MMC API, and whatever else happens out of it must
be managed out of it.