media: msm: camera_v2: Optimize VCM2 actuator code
It is useless and time consuming to allocate and deallocate the
configuration table everytime we need to set the actuator position.
Optimize the code by initializing the default parameters at VCM2
actuator init time, modifying only the required bits to set its
position when required and freeing the table when closing
specifically the VCM2.