clk: qcom: dsi-pll-28hpm: Implement hack for DBA-HDMI pixclock on Blanc
DBA-HDMI panels require this nasty hack to work because of some
miscalculation of the pixel clock (internally, not in Linux!) for
unknown reasons, and this is seen on at least APQ8056 SoMC Blanc.

Make them happy by fixing up the digital divider function in
run time with a new DT property.

Signed-off-by: Angelo G. Del Regno <kholk11@gmail.com>