clk: qcom: cpu-8996: Add dynamic speed bin (dynamic opp) support
MSM8996 has various versions with various speed bins and we
select them through a version number programmed in the specific
efuse that contains speed bin version and pvs version.

Add support to dynamically read the speed bin from DT and apply
it on the fly after reading the efuse and selecting the right one.