arm64: DT: msm-audio: Add missing DAIs for SDM660 internal codec
The SDM660 internal codec driver got updated in k4.9 and now
requires the additional quinary auxpcm and a mi2s4 node.
Add them even if they are used only in the SDM670 version of
this codec: they won't be used at all on 660 and 630 due to
the userspace configuration.
This avoids placing ifdefs in the audio driver.