clk: qcom: clk-hfpll: Notify the user only once if PLL lock failed
There is no reason to notify the user twice or to create a false
notification if later we succeed, so, we either send a message
to notify that the PLL locked when we retry to, or we just send
one when it really couldn't grab a lock.
The latter is the most convenient one.

Also, while at it, simplify the print.